The icon for the application.
I have used a closed source software called Aseprite to create the icon. Before their is no icon but their should now be an icon included. 

I will need to update the install file to allow the user to download the app to the app menu which will come with the icon. :)